added comments, added loading state
